HSA-backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
behave==1.2.7.dev6
behave-django==1.5.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.8.0
cryptography==44.0.3
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.1.2
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-sslserver==0.22
djangorestframework==3.16.0
fonttools==4.57.0
fpdf2==2.8.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
more-itertools==10.7.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.4
pillow==11.2.1
pip-chill==1.0.3
platformdirs==4.3.7
psycopg==3.2.7
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
selenium==4.32.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
django-rest-passwordreset==1.5.0
django-ratelimit==4.1.0
boto3==1.38.8